Serve BF16; retire the FP16-safety program
The value model's activations grow past FP16's range during training and
NaN when served in FP16 (the face-up-official ep4414 incident). The
overflow is a range problem, so serving BF16 -- FP32's 8-bit exponent --
removes it by construction, at a mantissa cost measured 20-50x below the
model's own Monte-Carlo error (quality metrics unchanged to 3 sig figs).

This lets the entire model-side containment program go: the FP32-pin
machinery, the FP16 export gate, and the activation-magnitude recipe
terms (PoolFcPenalty/lambda_pool_act and wld_z_loss/lambda_wld_z) across
all three trainer families. SimRunner's non-finite-leaf hard error stays
as the belt-and-suspenders tripwire.

Engine: Precision::kBF16 + BuilderFlag::kBF16; leaf service and
NeuralServiceOptions default to BF16; parity tests run BF16 alongside
FP16 against the FP32 reference. docs/fp16_safe_serving.md rewritten as a
resolved record.

+++ engine/include/agent/neural_service_options.h
@@ -19,7 +19,10 @@ struct NeuralServiceOptions {
   std::string model;
   int batch_size = 256;
   int cuda_device = 0;
-  std::string precision = "FP16";
+  // BF16 by default: its FP32-range exponent serves any checkpoint's activation
+  // magnitudes without FP16 overflow, at a mantissa cost well below the model's
+  // own error. FP16 remains selectable for models known to fit its range.
+  std::string precision = "BF16";
 
   // Register --model/--batch-size/--cuda-device/--precision, bound to this
   // struct's fields. Call before parsing the agent's option tokens; an agent
